requirements.txt
jira==3.5.2
python-dotenv==1.0.0
streamlit==1.28.2
azure-identity
azure-keyvault-secrets
azure-functions